Set up TravisCI and Coveralls.io. Also updated tests that broke. Should have done in separate commits - my bad.
