+++ .travis.yml
@@ -1,0 +1,31 @@
+# Config file for https://travis-ci.org/
+sudo: false
+
+language: node_js
+
+# Define the js versions against we want to test our code
+node_js:
+  - '8'
+
+before_install:
+  - npm update
+
+install:
+  # Install npm packages
+  - npm install
+
+# Testing the app (see phpunit.xml) for configs, generating Code Coverage report
+script:
+  - npm test
+  - npm run coveralls
+# see: https://stackoverflow.com/questions/22361462/coveralls-io-with-travis-ci
+
+# Tell Travis CI to monitor only 'main' branch
+branches:
+  only: testing
+
+# You can delete the cache using travis-ci web interface
+cache:
+  directories:
+    - node_modules
+    - $HOME/.sonar/cache
