--- client/src/components/RatingsAndReviews.jsx
@@ -1,8 +1,0 @@
-import React from 'react'
-
-function RatingsAndReviews (props) {
-
-  return (<div>Ratings and Reviews</div>)
-}
-
-export default RatingsAndReviews;
