+++ client/src/components/productOverview/ProductOverview.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const ProductOverview = (props) => {
+  return (
+    <div>Product Overview</div>
+  );
+};
+
+export default ProductOverview;
