+++ client/src/components/ratingsReviews/RatingsAndReviews.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const RatingsAndReviews = (props) => {
+  return (
+    <div>Ratings and Reviews</div>
+  );
+};
+
+export default RatingsAndReviews;
