+++ sonar-project.properties
@@ -1,0 +1,23 @@
+sonar.projectKey=continuous-integration-js
+sonar.projectName=continuous-integration-js
+sonar.projectVersion=1.0
+
+# =====================================================
+#   Properties that will be shared amongst all modules
+# =====================================================
+
+sonar.host.url=https://sonarcloud.io
+sonar.organization=markpthomas-github
+sonar.login=0570a973a952ba42a93f8f8c46b7f2cb7b6675cf
+
+# SQ standard properties
+sonar.sources=./src
+sonar.tests=./tests
+
+# Properties specific to language plugins:
+# Seems that sonar no longer includes detailed unit test info:
+# https://docs.sonarqube.org/display/PLUG/JavaScript+Coverage+Results+Import
+# https://stackoverflow.com/questions/15770877/how-to-tell-sonar-to-use-my-surefire-reports-for-unit-test-results
+# https://blog.akquinet.de/2014/11/25/js-test-coverage/
+# maybe, but a lot of work, maybe N/A: https://github.com/tornaia/karma-sonarqube-unit-reporter
+sonar.javascript.lcov.reportPaths=./coverage/reports/lcov.info
